feat: adicionar imagem

+++ src/main/java/Lokei/aplication/infrastructure/persistence/repository/AluguelEntityRepository.java
@@ -3,5 +3,5 @@ package Lokei.aplication.infrastructure.persistence.repository;
 import Lokei.aplication.infrastructure.persistence.entity.AluguelEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AluguelEntityRepository extends JpaRepository<AluguelEntity, Integer> {
+public interface AluguelEntityRepository extends JpaRepository<AluguelEntity, Long> {
 }
