+++ src/main/java/Lokei/aplication/infrastructure/persistence/repository/AvaliacaoEntityRepository.java
@@ -3,5 +3,5 @@ package Lokei.aplication.infrastructure.persistence.repository;
 import Lokei.aplication.infrastructure.persistence.entity.AvaliacaoEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AvaliacaoEntityRepository extends JpaRepository<AvaliacaoEntity, Integer> {
+public interface AvaliacaoEntityRepository extends JpaRepository<AvaliacaoEntity, Long> {
 }
