+++ src/main/java/Lokei/aplication/infrastructure/persistence/repository/DenunciaEntityRepository.java
@@ -3,5 +3,5 @@ package Lokei.aplication.infrastructure.persistence.repository;
 import Lokei.aplication.infrastructure.persistence.entity.DenunciaEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DenunciaEntityRepository extends JpaRepository<DenunciaEntity, Integer> {
+public interface DenunciaEntityRepository extends JpaRepository<DenunciaEntity, Long> {
 }
