+++ src/main/java/Lokei/aplication/infrastructure/persistence/repository/EnderecoEntityRepository.java
@@ -3,7 +3,7 @@ package Lokei.aplication.infrastructure.persistence.repository;
 import Lokei.aplication.infrastructure.persistence.entity.EnderecoEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EnderecoEntityRepository extends JpaRepository<EnderecoEntity, Integer> {
+public interface EnderecoEntityRepository extends JpaRepository<EnderecoEntity, Long> {
 }
