+++ src/main/java/Lokei/aplication/infrastructure/persistence/repository/ImagemEntityRepository.java
@@ -3,7 +3,7 @@ package Lokei.aplication.infrastructure.persistence.repository;
 import Lokei.aplication.infrastructure.persistence.entity.ImagemEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ImagemEntityRepository extends JpaRepository<ImagemEntity, Integer> {
+public interface ImagemEntityRepository extends JpaRepository<ImagemEntity, Long> {
 }
