+++ src/main/java/Lokei/aplication/infrastructure/persistence/repository/UsuarioEntityRepository.java
@@ -3,5 +3,5 @@ package Lokei.aplication.infrastructure.persistence.repository;
 import Lokei.aplication.infrastructure.persistence.entity.UsuarioEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UsuarioEntityRepository extends JpaRepository<UsuarioEntity, Integer> {
+public interface UsuarioEntityRepository extends JpaRepository<UsuarioEntity, Long> {
 }
